requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-model-utils
